dataService/requirements.txt
flask
firebase-admin
mysql-connector-python